requirements.txt
# Core dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
cerebras_cloud_sdk==1.50.1
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
firecrawl==4.3.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.6.4
nest-asyncio==1.6.0
newsapi-python==0.2.7
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
typing_extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1

# Note: pip and setuptools are typically pre-installed with Python
# pip==25.2
# setuptools==65.5.0